fix: imports, renamings and natspecs
