+++ packages/data-flow/src/registries/index.ts
@@ -1,3 +1,1 @@
-export * from "./cachedStrategyRegistry.js";
-export * from "./dbStrategyRegistry.js";
-export * from "./strategyRegistry.js";
+export * from "./strategy/index.js";
