+++ packages/data-flow/src/registries/strategy/index.ts
@@ -1,0 +1,2 @@
+export * from "./cachedStrategyRegistry.js";
+export * from "./dbStrategyRegistry.js";
